requirements.txt
django
git+https://github.com/zxh326/django-simple-serializer.git
image
Pillow
PyJWT
redis
celery
